+++ types/Parcel/ParcelCreateParameters.d.ts
@@ -1,0 +1,4 @@
+import { ParametersToOmitOnCreate } from '../utils';
+import { IParcel } from './Parcel';
+
+export declare interface IParcelCreateParameters extends Omit<IParcel, ParametersToOmitOnCreate> {}
