+++ types/Parcel/index.d.ts
@@ -1,1 +1,2 @@
 export * from './PredefinedPackage';
+export * from './Parcel';
